+++ internal/common/common.go
@@ -28,6 +28,8 @@ var ForceHTTPS = true
 const (
 	sessionChars       = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 	sessionTokenLength = 20
+	pairingCodeChars   = "0123456789"
+	pairingCodeLength  = 4
 )
 
 // AssertPathExists returns nil only if it has been successfully
@@ -274,15 +276,23 @@ type SSECtx struct {
 }
 
 func NewSessionToken() string {
-	r := make([]byte, sessionTokenLength)
+	return newRandomString(sessionTokenLength, sessionChars)
+}
+
+func NewPairingCode() string {
+	return newRandomString(pairingCodeLength, pairingCodeChars)
+}
+
+func newRandomString(count int, characterSet string) string {
+	r := make([]byte, count)
 	_, err := rand.Read(r)
 	if err != nil {
 		panic(err)
 	}
 
-	b := make([]byte, sessionTokenLength)
+	b := make([]byte, count)
 	for i := range b {
-		b[i] = sessionChars[r[i]%byte(len(sessionChars))]
+		b[i] = characterSet[r[i]%byte(len(characterSet))]
 	}
 	return string(b)
 }