+++ irmaclient/handlers.go
@@ -53,7 +53,7 @@ func (h *keyshareEnrollmentHandler) fail(err error) {
 }
 
 // Not interested, ingore
-func (h *keyshareEnrollmentHandler) StatusUpdate(action irma.Action, status irma.Status) {}
+func (h *keyshareEnrollmentHandler) StatusUpdate(action irma.Action, status irma.ClientStatus) {}
 
 // The methods below should never be called, so we let each of them fail the session
 func (h *keyshareEnrollmentHandler) RequestVerificationPermission(request *irma.DisclosureRequest, satisfiable bool, candidates [][]DisclosureCandidates, ServerName *irma.RequestorInfo, callback PermissionHandler) {
@@ -82,4 +82,7 @@ func (h *keyshareEnrollmentHandler) KeyshareEnrollmentMissing(manager irma.Schem
 }
 func (h *keyshareEnrollmentHandler) ClientReturnURLSet(clientReturnURL string) {
 	h.fail(errors.New("Keyshare enrollment session unexpectedly found an external return url"))
+}
+func (h *keyshareEnrollmentHandler) PairingRequired(pairingCode string) {
+	h.fail(errors.New("Keyshare enrollment session failed: device pairing required"))
 }