+++ server/irmac/irmac.go
@@ -12,6 +12,8 @@ import (
 	"net/http"
 	"net/http/httptest"
 
+	irma "github.com/privacybydesign/irmago"
+
 	"github.com/privacybydesign/irmago/server"
 	"github.com/privacybydesign/irmago/server/irmaserver"
 )
@@ -49,9 +51,10 @@ func Initialize(IrmaConfiguration *C.char) *C.char {
 func StartSession(requestString *C.char) (r *C.char) {
 	// Create struct for return information
 	result := struct {
-		IrmaQr string
-		Token  string
-		Error  string
+		IrmaQr          string
+		RequestorToken  string
+		FrontendRequest *irma.FrontendSessionRequest
+		Error           string
 	}{}
 	defer func() {
 		j, _ := json.Marshal(result)
@@ -65,7 +68,7 @@ func StartSession(requestString *C.char) (r *C.char) {
 	}
 
 	// Run the actual core function
-	qr, token, err := s.StartSession(C.GoString(requestString), nil)
+	qr, requestorToken, frontendRequest, err := s.StartSession(C.GoString(requestString), nil)
 
 	// And properly return the result
 	if err != nil {
@@ -79,7 +82,8 @@ func StartSession(requestString *C.char) (r *C.char) {
 	}
 	// return actual results
 	result.IrmaQr = string(qrJson)
-	result.Token = token
+	result.RequestorToken = string(requestorToken)
+	result.FrontendRequest = frontendRequest
 	return
 }
 
@@ -91,7 +95,7 @@ func GetSessionResult(token *C.char) *C.char {
 	}
 
 	// Run the actual core function
-	result := s.GetSessionResult(C.GoString(token))
+	result := s.GetSessionResult(irma.RequestorToken(C.GoString(token)))
 
 	// And properly return results
 	if result == nil {
@@ -113,7 +117,7 @@ func GetRequest(token *C.char) *C.char {
 	}
 
 	// Run the core function
-	result := s.GetRequest(C.GoString(token))
+	result := s.GetRequest(irma.RequestorToken(C.GoString(token)))
 
 	// And properly return results
 	if result == nil {
@@ -135,7 +139,7 @@ func CancelSession(token *C.char) *C.char {
 	}
 
 	// Run the core function
-	err := s.CancelSession(C.GoString(token))
+	err := s.CancelSession(irma.RequestorToken(C.GoString(token)))
 
 	if err != nil {
 		return C.CString(err.Error())