+++ server/requestorserver/server.go
@@ -94,7 +94,7 @@ func (s *Server) startClientServer() error {
 
 func (s *Server) startServer(handler http.Handler, name, addr string, port int, tlsConf *tls.Config) error {
 	fulladdr := fmt.Sprintf("%s:%d", addr, port)
-	s.conf.Logger.Info(name, " listening at ", fulladdr)
+	s.conf.Logger.Info(name, " listening at ", fulladdr, s.conf.ApiPrefix)
 
 	serv := &http.Server{
 		Addr:      fulladdr,
@@ -152,11 +152,17 @@ var corsOptions = cors.Options{
 	AllowedMethods: []string{http.MethodGet, http.MethodPost, http.MethodDelete},
 }
 
+func (s *Server) prefixRouter(router *chi.Mux) (prefixedRouter *chi.Mux) {
+	prefixedRouter = chi.NewRouter()
+	prefixedRouter.Mount(s.conf.ApiPrefix, router)
+	return
+}
+
 func (s *Server) ClientHandler() http.Handler {
 	router := chi.NewRouter()
 	router.Use(cors.New(corsOptions).Handler)
 	s.attachClientEndpoints(router)
-	return router
+	return s.prefixRouter(router)
 }
 
 func (s *Server) attachClientEndpoints(router *chi.Mux) {
@@ -195,7 +201,7 @@ func (s *Server) Handler() http.Handler {
 		// Server routes
 		r.Route("/session", func(r chi.Router) {
 			r.Post("/", s.handleCreateSession)
-			r.Route("/{token}", func(r chi.Router) {
+			r.Route("/{requestorToken}", func(r chi.Router) {
 				r.Delete("/", s.handleDelete)
 				r.Get("/status", s.handleStatus)
 				r.Get("/statusevents", s.handleStatusEvents)
@@ -219,7 +225,7 @@ func (s *Server) Handler() http.Handler {
 		r.Post("/revocation", s.handleRevocation)
 	})
 
-	return router
+	return s.prefixRouter(router)
 }
 
 func (s *Server) StaticFilesHandler() http.Handler {
@@ -295,7 +301,7 @@ func (s *Server) handleRevocation(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleStatus(w http.ResponseWriter, r *http.Request) {
-	res := s.irmaserv.GetSessionResult(chi.URLParam(r, "token"))
+	res := s.irmaserv.GetSessionResult(irma.RequestorToken(chi.URLParam(r, "requestorToken")))
 	if res == nil {
 		server.WriteError(w, server.ErrorSessionUnknown, "")
 		return
@@ -304,13 +310,13 @@ func (s *Server) handleStatus(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleStatusEvents(w http.ResponseWriter, r *http.Request) {
-	token := chi.URLParam(r, "token")
-	s.conf.Logger.WithFields(logrus.Fields{"session": token}).Debug("new client subscribed to server sent events")
+	requestorToken := chi.URLParam(r, "requestorToken")
+	s.conf.Logger.WithFields(logrus.Fields{"session": requestorToken}).Debug("new client subscribed to server sent events")
 	r = r.WithContext(context.WithValue(r.Context(), "sse", common.SSECtx{
 		Component: server.ComponentSession,
-		Arg:       token,
+		Arg:       requestorToken,
 	}))
-	if err := s.irmaserv.SubscribeServerSentEvents(w, r, token, true); err != nil {
+	if err := s.irmaserv.SubscribeServerSentEvents(w, r, requestorToken, true); err != nil {
 		server.WriteResponse(w, nil, &irma.RemoteError{
 			Status:      server.ErrorUnsupported.Status,
 			ErrorName:   string(server.ErrorUnsupported.Type),
@@ -320,14 +326,14 @@ func (s *Server) handleStatusEvents(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleDelete(w http.ResponseWriter, r *http.Request) {
-	err := s.irmaserv.CancelSession(chi.URLParam(r, "token"))
+	err := s.irmaserv.CancelSession(irma.RequestorToken(chi.URLParam(r, "requestorToken")))
 	if err != nil {
 		server.WriteError(w, server.ErrorSessionUnknown, "")
 	}
 }
 
 func (s *Server) handleResult(w http.ResponseWriter, r *http.Request) {
-	res := s.irmaserv.GetSessionResult(chi.URLParam(r, "token"))
+	res := s.irmaserv.GetSessionResult(irma.RequestorToken(chi.URLParam(r, "requestorToken")))
 	if res == nil {
 		server.WriteError(w, server.ErrorSessionUnknown, "")
 		return
@@ -346,8 +352,8 @@ func (s *Server) handleJwtResult(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	sessiontoken := chi.URLParam(r, "token")
-	res := s.irmaserv.GetSessionResult(sessiontoken)
+	requestorToken := irma.RequestorToken(chi.URLParam(r, "requestorToken"))
+	res := s.irmaserv.GetSessionResult(requestorToken)
 	if res == nil {
 		server.WriteError(w, server.ErrorSessionUnknown, "")
 		return
@@ -374,8 +380,8 @@ func (s *Server) handleJwtProofs(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	sessiontoken := chi.URLParam(r, "token")
-	res := s.irmaserv.GetSessionResult(sessiontoken)
+	requestorToken := irma.RequestorToken(chi.URLParam(r, "requestorToken"))
+	res := s.irmaserv.GetSessionResult(requestorToken)
 	if res == nil {
 		server.WriteError(w, server.ErrorSessionUnknown, "")
 		return
@@ -400,7 +406,7 @@ func (s *Server) handleJwtProofs(w http.ResponseWriter, r *http.Request) {
 		claims["iss"] = s.conf.JwtIssuer
 	}
 	claims["status"] = res.ProofStatus
-	validity := s.irmaserv.GetRequest(sessiontoken).Base().ResultJwtValidity
+	validity := s.irmaserv.GetRequest(requestorToken).Base().ResultJwtValidity
 	if validity != 0 {
 		claims["exp"] = time.Now().Unix() + int64(validity)
 	}
@@ -505,15 +511,16 @@ func (s *Server) createSession(w http.ResponseWriter, requestor string, rrequest
 	}
 
 	// Everything is authenticated and parsed, we're good to go!
-	qr, token, err := s.irmaserv.StartSession(rrequest, s.doResultCallback)
+	qr, requestorToken, frontendRequest, err := s.irmaserv.StartSession(rrequest, s.doResultCallback)
 	if err != nil {
 		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
 		return
 	}
 
 	server.WriteJson(w, server.SessionPackage{
-		SessionPtr: qr,
-		Token:      token,
+		SessionPtr:      qr,
+		Token:           requestorToken,
+		FrontendRequest: frontendRequest,
 	})
 }